move lowpass gain outside lowpass/lockin

+++ dsp/src/lockin.rs
@@ -8,8 +8,7 @@ pub struct Lockin {
 
 impl Lockin {
     /// Update the lockin with a sample taken at a given phase.
-    /// The lowpass has a gain of `1 << k`.
-    pub fn update(&mut self, sample: i16, phase: i32, k: u8) -> Complex<i32> {
+    pub fn update(&mut self, sample: i32, phase: i32, k: u8) -> Complex<i32> {
         // Get the LO signal for demodulation and mix the sample;
         let mix = Complex::from_angle(phase).mul_scaled(sample);
